feat: update semantics of write mask
not entirely sure what is the correctly semantics of mask port in
seq.mem. by looking at the generated verilog for rocketchip, its sram
has data width 64 and is provided with a 8 bit write mask. each of the
mask bit corresponds to the mask for a byte.

maybe this change is specifically tailored for rocket-chip

Signed-off-by: Zihan <zihanli0822@gmail.com>